+++ src/internal/grpcutil/bypass.go
@@ -1,4 +1,4 @@
-package internal
+package grpcutil
 
 import (
 	"bytes"
