+++ src/server/pfs/server/transaction_access.go
@@ -4,13 +4,11 @@ import (
 	"bytes"
 	"context"
 
-	"github.com/pachyderm/pachyderm/v2/src/pps"
-
+	"github.com/pachyderm/pachyderm/v2/src/internal/grpcutil"
 	"github.com/pachyderm/pachyderm/v2/src/internal/storage/fileset"
-
-	"github.com/pachyderm/pachyderm/v2/src/internal"
 	"github.com/pachyderm/pachyderm/v2/src/internal/transactionenv/txncontext"
 	"github.com/pachyderm/pachyderm/v2/src/pfs"
+	"github.com/pachyderm/pachyderm/v2/src/pps"
 )
 
 type FileAccessor struct {
@@ -28,6 +26,8 @@ func (a *apiServer) NewFileAccessor(ctx context.Context) txncontext.FileAccessor
 }
 
 // CreateFileset creates a fileset with a single file
+// TODO(peter): make this immediately exit the transaction and lead to extra transaction-wide
+// fileset handling logic, including renewal
 func (a *FileAccessor) CreateFileset(path string, data []byte, append bool) (string, error) {
 	id, err := a.d.createFileSet(a.ctx, func(uw *fileset.UnorderedWriter) error {
 		return uw.Put(path, "", append, bytes.NewReader(data))
@@ -39,5 +39,5 @@ func (a *FileAccessor) CreateFileset(path string, data []byte, append bool) (str
 }
 
 func (a *FileAccessor) GetPipelineDetails(info *pps.PipelineInfo) error {
-	return internal.GetPipelineDetails(a.ctx, a.apiServer, info)
+	return grpcutil.GetPipelineDetails(a.ctx, a.apiServer, info)
 }
